Bug fix Circular Average run out buffer
When calculating the circular average, I created a buffer wave to sum the intensity per pixel. The size of the buffer wave is automatically adjusted using a simple algorithm. However, this algorithm does not work well when the detector pixel is very fine and the detector size is very large. I increased the buffer size by four times and have now fixed this bug.